+++ src/interfaces/ICalculator.sol
@@ -5,5 +5,5 @@ interface ICalculator {
     function initCalculator() external;
     function calculateFee(uint256) external returns (uint256);
     function getFeePercentage() external view returns (uint256);
-    function updateFeePercentage(uint256, bytes memory message, bytes[] memory signatures) external;
+    function updateFeePercentage(uint248, bytes memory message, bytes[] memory signatures) external;
 }
