+++ src/libs/LibCalculator.sol
@@ -4,9 +4,10 @@ pragma solidity 0.8.23;
 library LibCalculator {
     bytes32 constant STORAGE_SLOT = keccak256("calculator.storage");
 
+    // one storage slot
     struct Storage {
         bool initialized;
-        uint256 feePercentage; // 5% = 500
+        uint248 feePercentage; // 5% = 500
     }
 
     function getCalculatorStorage() internal pure returns (Storage storage ds) {
